requirements.txt
asgiref==3.2.3
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-guardian==2.2.0
djangorestframework==3.11.0
drf-jwt==1.15.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
